requirements.txt
pandas~=2.1.0
flask~=2.3.3
seaborn~=0.12.2
matplotlib~=3.7.2
spacy~=3.6.1
numpy~=1.25.2
transformers~=4.30.2
scikit-learn~=1.3.0
sentence-transformers~=2.2.2
faiss-cpu~=1.7.4